fix(schemas): serialize order/sale money fields as Decimal, not float (M-8)
A Decimal DB value serialized through a float-typed Pydantic field silently
reintroduces binary float imprecision at the API boundary, defeating the
Decimal money-math fix upstream. Convert the response fields that trace
back to the columns/computations touched in this fix:
- checkout.py: CheckoutOrderItemResponse.unit_price/line_total,
  CheckoutOrderResponse.total_amount
- order.py: OrderItemResponse.price/line_total, OrderResponse.total_amount
- admin/sales.py: SaleResponse.total_amount/final_price/cost_price/profit

Deliberately left as float (out of scope for this fix): CheckoutSummaryResponse's
subtotal/tax/grand_total (computed in urjaa-backend's own route code — a
separate task); CartItemResponse.unit_price/line_total and
Product*.starting_price (display-only, never persisted — see PR notes for
a flagged follow-up); BulkSaleResponse and Sale*Request.final_price
(manual/POS sale entry, a different computation path from the checkout/
cart pricing chain this fix addresses); AdminVariantResponse.cost_price
and Variant*Request.cost_price (raw admin-entered input/passthrough, not
computed by this fix); dashboard/analytics aggregates in management.py
(out of scope — a broader, separate area).

+++ urjaa_core/schemas/order.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 from typing import Literal
 from uuid import UUID
 
@@ -31,8 +32,8 @@ class OrderItemResponse(BaseModel):
     product_name: str
     product_image: str | None
     quantity: int
-    price: float
-    line_total: float
+    price: Decimal
+    line_total: Decimal
 
     class Config:
         orm_mode = True
@@ -46,7 +47,7 @@ class OrderResponse(BaseModel):
     full_name: str
     phone: str | None
     shipping_address: OrderShippingAddress
-    total_amount: float
+    total_amount: Decimal
     currency: str
     payment_status: str
     formatted_total: str
